+++ package.json
@@ -5,7 +5,7 @@
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
+    "start": "nodemon --exec babel-node src/server.js"
   },
   "repository": {
     "type": "git",
@@ -22,6 +22,12 @@
   },
   "homepage": "https://github.com/ThaiBangHOANG/basic-nodeJS#readme",
   "dependencies": {
-    "express": "4.17.1"
+    "@babel/core": "7.15.5",
+    "@babel/node": "7.15.4",
+    "@babel/preset-env": "7.15.6",
+    "body-parser": "1.19.0",
+    "ejs": "3.1.6",
+    "express": "4.17.1",
+    "nodemon": "2.0.12"
   }
 }
