--- server.js
@@ -1,11 +1,0 @@
-const express = require("express");
-const app = express();
-const port = 3000;
-
-app.get("/", (req, res) => {
-  res.send("Hello from Thai Bang");
-});
-
-app.listen(port, () => {
-  console.log(`First Programm of Node JS at url: http://localhost:${port}`);
-});
